Add Protocols model, maintaince issue
move MedicineDose model to pharmacy app from main
